+++ CHANGES.rst
@@ -4,6 +4,15 @@
 Release Notes
 =============
 
+2.0.1 (Unreleased)
+==================
+
+- Update ``utils.calc_pixmap`` code to be ready for upcoming changes in GWCS
+  due to which inverse WCS transformations will respect bounding box by
+  allowing the caller of ``utils.calc_pixmap`` to disable the bounding box(es)
+  on both or one of the input WCS when computing the pixel map. [#164]
+
+
 2.0.0 (2024-10-23)
 ==================